+++ src/service-token/browser.js
@@ -4,7 +4,7 @@ const Config = require('../config')
 const Cache = require('../cache')
 const Deserializer = require('../jsonapi-serializer/deserializer')
 const { isString } = require('../utils/typecheck')
-const { simpleFetch } = require('../utils/simple-fetch')
+const simpleFetch = require('../utils/simple-fetch')
 
 const deserializer = new Deserializer({ keyForAttribute: 'camelCase' })
