+++ src/utils/simple-fetch.js
@@ -2,7 +2,7 @@
 
 const { fetch, Headers } = require('./shark-fetch')
 const Logger = require('../logger')
-const Error = require('../jsonapi-serializer/error')
+const { jsonApiError } = require('./response-helper')
 
 /**
  * Parses the JSON returned by a network request.
@@ -88,33 +88,4 @@ function simpleFetch (url, options) {
   })
 }
 
-function jsonApiError (response) {
-  const json = response.json
-  let errors = []
-
-  if (Array.isArray(json.errors)) {
-    errors = json.errors
-  } else {
-    // This handles "legacy" errors from our services that haven't changed to
-    // JSONAPI-compliant errors yet.
-    let errorDetails = []
-    if (json.messages) {
-      errorDetails = json.messages
-    } else if (json.message) {
-      errorDetails = [json.message]
-    } else {
-      Logger.log('Unhandled response type: ', response)
-      errorDetails = ['Unhandled error']
-    }
-    errors = errorDetails.map(detail => {
-      return { status: response.status, title: response.statusText, detail: detail }
-    })
-  }
-
-  return new Error(errors)
-}
-
-module.exports = {
-  simpleFetch,
-  jsonApiError
-}
+module.exports = simpleFetch
